% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{get_weekly_timeseries}
\alias{get_weekly_timeseries}
\title{Download and process multiple years of data}
\usage{
get_weekly_timeseries(disease = "Noroviral gastroenteritis", years,
  region_level = "State")
}
\arguments{
\item{disease}{which disease to retrieve data for}

\item{years}{which years to retrieve data for}

\item{region_level}{which level to retrieve data for}
}
\value{
a tibble with data
}
\description{
Download and process multiple years of data
}
