% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{get_weekly_timeseries_one_yr}
\alias{get_weekly_timeseries_one_yr}
\title{Download and format a weekly timeseries from one year to data.frame}
\usage{
get_weekly_timeseries_one_yr(disease = "Noroviral gastroenteritis",
  year = "2019", region_level = "State")
}
\arguments{
\item{disease}{which disease to retrieve data for}

\item{year}{which year to retrieve data for}

\item{region_level}{which level to retrieve data for}
}
\value{
a tibble with data
}
\description{
API only allows one year downloaded at a time (max rows per query).
So this function helps split queries up.
}
