% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{getHierarchies}
\alias{getHierarchies}
\title{More utilities}
\usage{
getHierarchies(cube, language)
}
\arguments{
\item{cube}{}

\item{language}{}
}
\value{
something XMLish
}
\description{
More utilities
}
