% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{getHierarchyMembers}
\alias{getHierarchyMembers}
\title{Utilities for accessing data}
\usage{
getHierarchyMembers(cube, language, filter)
}
\arguments{
\item{cube}{}

\item{language}{}

\item{filter}{}
}
\value{
something XMLish
}
\description{
Utilities for accessing data
}
