% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{getOlapData}
\alias{getOlapData}
\title{more utilities}
\usage{
getOlapData(cube, language, hierarchy, facet, filter, filterValue, filter2,
  filterValue2)
}
\arguments{
\item{cube}{}

\item{language}{}

\item{hierarchy}{}

\item{facet}{}

\item{filter}{}

\item{filterValue}{}

\item{filter2}{}

\item{filterValue2}{}
}
\value{
a data.frame
}
\description{
more utilities
}
