% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survstat_webservice.R
\name{getXMLFromWebService}
\alias{getXMLFromWebService}
\title{method to recieve the xml from the WebService}
\usage{
getXMLFromWebService(body_, service_)
}
\arguments{
\item{body_}{the xml request}

\item{service_}{the WebService method}
}
\value{
something XMLish
}
\description{
method to recieve the xml from the WebService
}
